Separate meeting snapshot analysis
Move participant detection, chat interpretation, and composer targeting into a focused macOS analysis module while preserving the public API and accessibility behavior.
